requirements.txt
langchain
langchain-openai
langchain-community
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
torch
# uncomment to use huggingface llms
huggingface-hub

# uncomment to use instructor embeddings
InstructorEmbedding
sentence-transformers
